Enable persistent storage of sandy db in docker
Use a wrap on sandy which checkes if the volume /sandy/db is mounted and
if there is a database db.sqlite3 inside it. If so, the wrapper symlinks
the user databse where sandy expects to; if not, the wrapper moves the
sandy database to /sandy/db and symlinks it back.